Remove dependency on github.com/microsoftgraph/msgraph-sdk-go
We only use msgraph-sdk-go for a handful of calls, and they are
relatively straightforward HTTP calls.

This code still uses the Microsoft packages for authentication, which
is the trickier part.

I'm having trouble getting the tests to pass consistently, but
I always have trouble with these tests.

I compiled the Vault binary with this change only, and it reduces
the size of the Vault binary from 360 MB to 264 MB.